Change bankDirection back to f16
A full f32 is unneeded for this operation.

+++ inc/player.h
@@ -38,7 +38,7 @@ typedef struct _Stage Stage;
 struct _Player {
   V2f32 position;
   Sprite* sprite;
-  f32 bankDirection;
+  f16 bankDirection;
   f32 attackCooldown;
   f32 damageCooldown;
   u8 health;
